Move exemplar comparator to exemplar_util.
